+++ einvoice_lens/__init__.py
@@ -8,4 +8,4 @@ __all__ = [
     "parse_commerical_invoice",
 ]
 
-__version__ = "0.3.2"
+__version__ = "0.3.4"